added css and refactored to use react router
